requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file test.txt test.in
#
alembic==0.9.6            # via flask-migrate
amqp==2.2.2               # via kombu
appnope==0.1.0            # via ipython
argparse==1.4.0           # via scoop
attrs==17.3.0             # via pytest
babel==2.5.1              # via flask-babelex
bcrypt==3.1.4
billiard==3.5.0.3         # via celery
bleach==2.1.1             # via nbconvert
blinker==1.4              # via flask-mail, flask-principal
bokeh==0.12.11
celery==4.1.0
cffi==1.11.2              # via bcrypt
click==6.7                # via flask, pip-tools
cycler==0.10.0            # via matplotlib
deap==1.2.2
decorator==4.1.2          # via ipython, traitlets
entrypoints==0.2.3        # via nbconvert
first==2.0.1              # via pip-tools
flask-admin==1.5.0
flask-babelex==0.9.3      # via flask-security
flask-login==0.4.1        # via flask-security
flask-mail==0.9.1
flask-migrate==2.1.1
flask-principal==0.4.0    # via flask-security
flask-security==3.0.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==0.12.2
greenlet==0.4.12          # via scoop
html5lib==1.0b10          # via bleach
ipykernel==4.6.1          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==6.2.1            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.0.5         # via jupyter
itsdangerous==0.24        # via flask, flask-security
jedi==0.11.0              # via ipython
jinja2==2.10              # via bokeh, flask, flask-babelex, nbconvert, notebook
jsonschema==2.6.0         # via nbformat
jupyter-client==5.1.0     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.2.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
kombu==4.1.0              # via celery
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
matplotlib==2.1.0
mistune==0.8.1            # via nbconvert
nbconvert==5.3.1          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.2.2           # via jupyter, widgetsnbextension
numpy==1.13.3
pandas==0.21.0
pandocfilters==1.4.2      # via nbconvert
parso==0.1.0              # via jedi
passlib==1.7.1            # via flask-security
pexpect==4.3.0            # via ipython
pickleshare==0.7.4        # via ipython
pint==0.8.1
pip-tools==1.11.0
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython, jupyter-console
psycopg2==2.7.3.2
ptyprocess==0.5.2         # via pexpect, terminado
py==1.5.2                 # via pytest
pycparser==2.18           # via cffi
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.2.0          # via matplotlib
pytest==3.3.0
python-dateutil==2.6.1    # via alembic, bokeh, jupyter-client, matplotlib, pandas
python-editor==1.0.3      # via alembic
pytz==2017.3              # via babel, celery, matplotlib, pandas
pyyaml==3.12
pyzmq==16.0.3             # via jupyter-client, scoop
qtconsole==4.3.1          # via jupyter
scoop==0.7.1.1
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via bcrypt, bleach, bokeh, cycler, html5lib, matplotlib, pip-tools, prompt-toolkit, pytest, python-dateutil, traitlets
speaklater==1.3           # via flask-babelex
sqlalchemy==1.1.15        # via alembic, flask-sqlalchemy
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
tornado==4.5.2            # via bokeh, ipykernel, notebook, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
vine==1.1.4               # via amqp
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
werkzeug==0.12.2          # via flask
widgetsnbextension==3.0.8  # via ipywidgets
wtforms==2.1              # via flask-admin, flask-wtf
